로드맵 API 구현

+++ src/main/java/com/consoltant/consoltant/domain/bestroadmap/dto/BestRoadmapResponseDto.java
@@ -13,7 +13,7 @@ public class BestRoadmapResponseDto {
 
     private User user;
 
-    private Long salary;
+    private Integer salary;
 
     private Long startAsset;
